Unique the list of handlers for help to display
Certain handlers (e.g. lita-locker) define routes in separate classes
under the same namespace, which caused them to be repeated in response
to the `help` command. This pull request solves the issue by uniquing
the list of handlers first.

I also discovered an issue where setting the bot's alias to '?' messed
with the regex in `filter_help`; I dealt with this by escaping the text
interpolated into the regex.