Improve help handler usability
Changes to the help handler made in #1 required that the user be
familiar with the internals of the bot, e.g. knowing which handler
namespace a particular command is defined under in order to find it.
This PR removes the `help HANDLER STRING` syntax and simplifies it to
`help STRING`, where the argument can represent either the name of a
handler or a substring on which to filter all commands.

Example of usage and output:
> !help log
Handler 'log' defines the following commands:
!log LEVEL MESSAGE - Log a message to Lita's Logger object. Valid levels: UNKNOWN, FATAL, ERROR, WARN, INFO, DEBUG

!locker log <label> - Show up to the last 10 activity log entries for <label>